+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `user:session:create` command.
 - Update dependencies.
 - Rework `database:drop` command.
+- Token management now uses Redis.
 
 ## 0.0.4 - 2023-04-22